# REST-API/cat/static/swagger.json
openapi: 3.0.3
info:
  title: CAT
  description: Case Assignment Tool
  termsOfService: http://swagger.io/terms/
  contact:
    email: himanshu.surendra@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: /
...
paths:
  /api/login:
    post:
      tags:
        - user
      summary: Login check
      description: ''
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid Input
        '401':
          description: Authentication Failure
...
  /api/user/{username}:
    get:
      tags:
        - user
      summary: Get user by user name
      description: ''
      operationId: getUserByName
      parameters:
        - name: username
          in: path
          description: The name that needs to be fetched.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    post:
      tags:
        - user
      summary: Update user
      description: ''
      operationId: updateUser
      parameters:
        - name: username
          in: path
          description: The user details that needs to be fetched.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditUser'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid Input
        '404':
          description: Object not found
        '401':
          description: Unauthorized Access
    delete:
      tags:
        - user
      summary: Delete user
      description: ''
      operationId: deleteUser
      parameters:
        - name: username
          in: path
          description: The user name that needs to be deleted.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid Input
        '404':
          description: Object not found
        '401':
          description: Unauthorized Access
  /api/user:
    post:
      tags:
        - user
      summary: Add a new user
      description: ''
      operationId: addUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUser'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid Input
  /api/user/reset-password/{email}:
    post:
      tags:
        - user
      summary: Reset user password
      description: ''
      operationId: UserResetPassword
      parameters:
        - name: email
          in: path
          description: The user's email whose password needs to be changed
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid Input
        '404':
          description: User not found
  /api/user/change-password/{username}:
    post:
      tags:
        - user
      summary: Change user password
      description: ''
      operationId: UserChangePassword
      parameters:
        - name: username
          in: path
          description: The username whose password needs to be changed
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePassword'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid Input
        '404':
          description: User not found
  /api/user/reactivate/{username}:
    post:
      tags:
        - user
      summary: Schedule User reactivation
      description: ''
      operationId: reactivate
      parameters:
        - name: username
          in: path
          description: The username that needs to be scheduled for reactivation
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReactivateUser'
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid Input
  /api/user/schedule-shift-change/{username}:
    post:
      tags:
        - user
      summary: Schedule User shift change
      description: ''
      operationId: scheduleshiftchange
      parameters:
        - name: username
          in: path
          description: The username whose shift change needs to be scheduled
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleShiftChange'
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid Input
  /api/user/schedule-handoffs/{username}:
    post:
      tags:
        - user
      summary: Schedule User cases handoffs
      description: ''
      operationId: schedulehandoffs
      parameters:
        - name: username
          in: path
          description: The username whose cases needs to be shcheduled for handoffs
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleHandoffs'
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid Input
  /api/delete-job/{jobid}:
    delete:
      tags:
        - user
      summary: Delete a scheduled job
      description: ''
      operationId: deleteJob
      parameters:
        - name: jobid
          in: path
          description: The jobid of job that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
        '404':
          description: Not Found
  /api/jobs/{page}:
    get:
      tags:
        - user
      summary: Get all jobs (audit)
      description: ''
      operationId: getJobs
      parameters:
        - name: page
          in: path
          description: The page number to be fetched.
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Jobs'
            application/xml:
              schema:
                $ref: '#/components/schemas/Jobs'
...
  /api/add-salesforce-email:
    post:
      tags:
        - user
      summary: Add a email template
      description: ''
      operationId: addSalesforceEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSalesforceEmail'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesforceEmail'
            application/xml:
              schema:
                $ref: '#/components/schemas/SalesforceEmail'
        '400':
          description: Invalid Input
  /api/salesforce-email/{email_name}:
    get:
      tags:
        - user
      summary: Get email template by template name
      description: ''
      operationId: getSalesforceEmail
      parameters:
        - name: email_name
          in: path
          description: The email templates name that needs to be fetched.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesforceEmail'
            application/xml:
              schema:
                $ref: '#/components/schemas/SalesforceEmail'
        '404':
          description: Email template not found
    post:
      tags:
        - user
      summary: Update email template
      description: ''
      operationId: UpdateSalesforceEmail
      parameters:
        - name: email_name
          in: path
          description: The email templates name that needs to be edited
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalesforceEmailModel'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesforceEmail'
            application/xml:
              schema:
                $ref: '#/components/schemas/SalesforceEmail'
        '400':
          description: Invalid Input
        '404':
          description: Object not found
        '401':
          description: Unauthorized Access
    delete:
      tags:
        - user
      summary: Delete an email template
      description: ''
      operationId: deleteSalesforceEmail
      parameters:
        - name: email_name
          in: path
          description: The email template user name that needs to be deleted.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesforceEmail'
            application/xml:
              schema:
                $ref: '#/components/schemas/SalesforceEmail'
        '400':
          description: Invalid Input
        '404':
          description: Object not found
        '401':
          description: Unauthorized Access
  /api/jinja2_variables:
    get:
      tags:
        - user
      summary: Get all j2 variables exposed to users
      description: ''
      operationId: getJinja2Variables
      responses:
        '200':
          description: successful operation
        '401':
          description: Unauthorized Access
...
components:
  schemas:
...
    UserModel:
      type: object
      properties:
        email:
          type: string
          example: abc@efg.com
        teamname:
          type: string
          example: Team1
        active:
          type: boolean
          description: default is 1
        admin:
          type: boolean
          description: whether the user is admin or not. Default is false
        team_email_notifications:
          type: boolean
          description: whether the user recives team activity notifications. Default is false
        monitor_case_updates:
          type: boolean
          description: Monitor users cases for external updates when user is ooo or on weekends. Default is false
        shift_start:
          type: string
          example: '09:00:00'
        shift_end:
          type: string
          example: '18:00:00'
        first_name:
          type: string
          example: 'Tom'
        last_name:
          type: string
          example: 'Kumar'
        timezone:
          type: string
          example: US/Pacific
          description: python lib pytz format
    User:
      allOf:
        - type: object
        - $ref: '#/components/schemas/UserModel'
      properties:
        username:
          type: string
          example: u1
        user_since:
          type: string
          description: default is current timestamp when adding a new user.
          example: Fri, 06 Oct 2023 01:58:56 GMT
        products:
          type: string
          example: Product1
        full_name:
          type: string
          example: Bob Marley
        last_login:
          type: string
          description: default is user creation timestamp. keeps track of users last login.
          example: Sat, 07 Oct 2023 12:45:56 GMT
        in_shift:
          type: boolean
          description: whether the user is in shift or not when the call is made
          example: true
    AddUser:
      allOf:
        - $ref: '#/components/schemas/UserModel'
        - type: object
      required:
        - username
        - password
        - teamname
      properties:
        username:
          type: string
          example: u1
        password:
          type: string
    EditUser:
      allOf:
        - $ref: '#/components/schemas/UserModel'
        - type: object
    ChangePassword:
      required:
        - current_password
        - new_password
      properties:
        current_password:
          type: string
          example: test
        new_password:
          type: string
          example: test123
...
    Login:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: User1
        password:
          type: string
          example: pass123
        muleuser:
          type: boolean
          example: False
...
    ReactivateUser:
      type: object
      required:
        - datetime
        - timezone
        - active
      properties:
        datetime:
          type: string
          example: 2023/10/19 13:12
        timezone:
          type: string
          example: US/Pacific
        active:
          type: boolean
    ScheduleShiftChange:
      type: object
      required:
        - datetime
        - timezone
        - shift_start
        - shift_end
      properties:
        datetime:
          type: string
          example: 2023/10/19 13:12
        timezone:
          type: string
          example: US/Pacific
        shift_start:
          type: string
          example: '09:00:00'
        shift_end:
          type: string
          example: '18:00:00'
    ScheduleHandoffs:
      type: object
      required:
        - datetime
        - timezone
      properties:
        datetime:
          type: string
          example: 2023/10/19 13:12
        timezone:
          type: string
          example: US/Pacific
        check_in_shift:
          type: boolean
          example: false
          description: whether to consider users shift hours when cat'ing the handoffs. Default is false
...
    Jobs:
      type: object
      properties:
        number:
          type: integer
        job_id:
          type: string
          example: deactivate_user_test
        username:
          type: string
          example: user1
          description: Username who submitted the job
        job_type:
          type: string
          example: reactivate_user
        details:
          type: string
          example: {"username": "bhimanshu", "datetime": "2023-11-08 11:22:00", "timezone": "US/Pacific", "active": false}
          description: A json formated string.
        time:
          type: string
          example: Mon, 02 Oct 2023 16:56:10 GMT
        status:
          type: string
          example: Success
    SalesforceEmailModel:
      type: object
      properties:
        email_body:
          type: string
          example: This is example email body
          description: The email body 
        email_subject:
          type: string
          example: This is example email subject
          description: The email subject 
    AddSalesforceEmail:
      type: object
      required:
        - email_name
        - email_body
      allOf:
        - $ref: '#/components/schemas/SalesforceEmailModel'
        - type: object
      properties:
        email_name:
          type: string
          example: email1
    SalesforceEmail:
      type: object
      allOf:
        - $ref: '#/components/schemas/SalesforceEmailModel'
        - type: object
      properties:
        id:
          type: integer
          example: 2
          description: internal id given to this email
        user:
          type: string
          example: user1
          description: Username of user who created the email
        email_name:
          type: string
          example: email1